Motor

+++ FreeRTOS_test/HARDWARE/MOTOR/motor.h
@@ -1,15 +1,15 @@
 /*********************************************************************************************************************
  * @file       		motor.h
  * @version    		V1.0
- * @Software 			IAR 8.3 or MDK 5.24
+ * @Software 		IAR 8.3 or MDK 5.24
  * @TargetCore		MK66FX
  * @date       		2019-12-13
- * @note					电机PWM
- *								Peripheral			MCU
- *								L_PWM1					PD4(TM0 ch4)
- *								L_PWM2					PD5(TM0 ch5)
- *								R_PWM1					PD6(TM0 ch6)
- *								R_PWM2					PD7(TM0 ch7)
+ * @note			电机PWM
+ *					Peripheral		MCU
+ *					L_PWM1			PD4(TM0 ch4)
+ *					L_PWM2			PD5(TM0 ch5)
+ *					R_PWM1			PD6(TM0 ch6)
+ *					R_PWM2			PD7(TM0 ch7)
  ********************************************************************************************************************/
  
 #ifndef __MOTOR_H
@@ -33,10 +33,10 @@
 
 typedef volatile struct	Motor_PWM_Type
 {
-	uint16_t right_pwm1;
-	uint16_t right_pwm2;
-	uint16_t left_pwm1;
-	uint16_t left_pwm2;
+	int right_pwm1;
+	int right_pwm2;
+	int left_pwm1;
+	int left_pwm2;
 } __Motor_PWM;
 
 extern __Motor_PWM Motor_PWM;
